fix: corrigido campos dos dtos.

+++ src/main/java/br/com/gestao_oficina_api/domain/dto/request/ServicoCreateDTO.java
@@ -1,8 +1,9 @@
 package br.com.gestao_oficina_api.domain.dto.request;
 
-import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
+
+import java.math.BigDecimal;
 
 public record ServicoCreateDTO (
 
@@ -10,7 +11,7 @@ public record ServicoCreateDTO (
         String descricao,
 
         @NotNull(message = "O preço base é obrigatório")
-        Double precoBase,
+        BigDecimal precoBase,
 
         @NotNull(message = "A duração média é obrigatória")
         Integer duracaoMedia
