+++ src/main/java/br/com/gestao_oficina_api/domain/dto/response/OrdemServicoResponseDTO.java
@@ -3,6 +3,7 @@ package br.com.gestao_oficina_api.domain.dto.response;
 import br.com.gestao_oficina_api.domain.model.Servico;
 import br.com.gestao_oficina_api.domain.model.Veiculo;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -16,7 +17,7 @@ public record OrdemServicoResponseDTO(
 
         String status,
 
-        Double valorEstimado,
+        BigDecimal valorEstimado,
 
         Veiculo veiculo,
