+++ src/main/java/br/com/gestao_oficina_api/domain/dto/response/ServicoResponseDTO.java
@@ -1,12 +1,14 @@
 package br.com.gestao_oficina_api.domain.dto.response;
 
+import java.math.BigDecimal;
+
 public record ServicoResponseDTO (
 
         Long id,
 
         String descricao,
 
-        Double precoBase,
+        BigDecimal precoBase,
 
         Integer duracaoMedia
 ) {}
